Refactor Lepiter and GT installer
Remove duplication that caused a prolbem when changing the support folders to install (forgot to do both classes).

Use a trait to allow each to inherit from unrelated ancestors. That's important because we want GT to recognize each as an RlResource so that the proper views are shown, but the way ResourcesLive determines what to show is a negotiation between a class and its subclasses. Thus, if GT installer was not a zip archive subclass, zip would be in a race condition with it.