Mac Launch Services Spike
Handles the simple case where there is no entry for file type. Seems the underlying support for this use case offered by Mac is extremely primitive. There used to be a CLI util called duti that we were hoping to wrap, but it doesn't seem to be updated for recent macOS. Won't work on this more for now.